Update policy.md
Two fairly small changes to provisions under Information Management and Access.  I cut the discussion about possible proprietary and information protection interests of contractors because it seemed unnecessary to try listing out the factors agencies may have to consider when trying to meet open data objectives for data that doesn't fall squarely in the public domain.  These may be some of the concerns now but in 5 years other factors may be more prominent.  Instead I think it makes more sense to keep it simple.  The provision already says "appropriate provisions" which gives agencies enough leeway.  
On the second provision, I again thought it was unnecessary to try to list where cost concerns derived from.  The original draft highlighted cost of dissemination and accessibility improvements.  Again these may be the source of concerns now but given that we are drafting for the next decade or two, it makes more sens to simply say "considers costs." 
The final words of the provision were confusing to me "and demonstrate usefulness of the information"  I couldn't tell if this was the third thing being listed that online access had to meet -- informed by the public, considers costs, and demonstrates the usefulness of the info.  Or if it was another aspect that access needed to be balanced against - costs and the demonstrated usefulness of info. The first approach would probably be acceptable but didn't seem vital and the second possibility was very troubling.  So I suggested cutting it.  If there is a good reason to keep it, it should be made much more clear.